GENAI/Langchain/Agents/code/requirements.txt
langchain-openai
langchain-community
langchain-core
requests
duckduckgo-search

langchain-ollama
langchain-huggingface
python-dotenv